+++ packages/orm/angel_orm_generator/example/main.dart
@@ -21,8 +21,12 @@ class _FakeExecutor extends QueryExecutor {
 
   @override
   Future<List<List>> query(
-      String tableName, String? query, Map<String, dynamic> substitutionValues,
-      [returningFields = const []]) async {
+    String tableName,
+    String? query,
+    Map<String, dynamic> substitutionValues, {
+    String returningQuery = '',
+    List<String> returningFields = const [],
+  }) async {
     var now = DateTime.now();
     print(
         '_FakeExecutor received query: $query and values: $substitutionValues');
@@ -35,6 +39,10 @@ class _FakeExecutor extends QueryExecutor {
   Future<T> transaction<T>(FutureOr<T> Function(QueryExecutor) f) {
     throw UnsupportedError('Transactions are not supported.');
   }
+
+  @override
+  // TODO: implement dialect
+  Dialect get dialect => PostgreSQLDialect();
 }
 
 @orm