+++ packages/orm/angel_orm_generator/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 5.0.0
+
+* Updated `analyzer` to version 3.x
+* Updated min SDK to 2.14.x
+
 ## 4.3.0
 
 * Added `where.raw()`